Resolve merge conflicts, enforce C++14 for GCC 5 builds, and perform code cleanups including type fixes and include adjustments.

+++ usr.sbin/mtree/verify.c
@@ -185,7 +185,7 @@ miss(NODE *p, char *tail)
 			continue;
 		if (p->type != F_DIR && (dflag || p->flags & F_VISIT))
 			continue;
-		if (snprintf(tail, sizeof(path) - (tail - path), "%s", p->name) >=
+		if ((size_t)snprintf(tail, sizeof(path) - (tail - path), "%s", p->name) >=
 		    sizeof(path) - (tail - path))
 			mtree_err("Pathname too long.");
 		if (!(p->flags & F_VISIT)) {
